Debounce signal whilst handling interrupt
This should be much bettererer. The previous debounce was happening once
per input check (8ms in main.rs), which is quite a long period of time
to debounce over. Now, the interrupts are handled once per input check
(8ms), and immediately debounced. The likelihood of needing debounce is
purely when the input check and the interrupt happen shortly after each
other.